Extract `apiGetResults` method
