+++ frontend/src/metabase/query_builder/actions/querying.js
@@ -6,6 +6,7 @@ import * as MetabaseAnalytics from "metabase/lib/analytics";
 import { startTimer } from "metabase/lib/performance";
 import { defer } from "metabase/lib/promise";
 import { createThunkAction } from "metabase/lib/redux";
+import { runQuestionQuery as apiRunQuestionQuery } from "metabase/services";
 
 import { getMetadata } from "metabase/selectors/metadata";
 import { getSensibleDisplays } from "metabase/visualizations";
@@ -124,12 +125,11 @@ export const runQuestionQuery = ({
 
     const queryTimer = startTimer();
 
-    question
-      .apiGetResults({
-        cancelDeferred: cancelQueryDeferred,
-        ignoreCache: ignoreCache,
-        isDirty: cardIsDirty,
-      })
+    apiRunQuestionQuery(question, {
+      cancelDeferred: cancelQueryDeferred,
+      ignoreCache: ignoreCache,
+      isDirty: cardIsDirty,
+    })
       .then(queryResults => {
         queryTimer(duration =>
           MetabaseAnalytics.trackStructEvent(
